[ci skip] Using `BOTO_CONFIG` in `populate_boto.py`.
This was the code can be used once, rather than having nearly
identical scripts in `travis/` and `appveyor/`.

This **infers** the build directory simply as the directory that
contains the `.boto` file and uses that directory to access
`key.json`.

Also adding comments to `.travis.yml` to make the `gsutil`
sections have identical steps and descriptions for those steps.